20100509-1512 - Replace BETA with PREVIEW for some labor firmwar, to be compatibel with freetz.
git-svn-id: https://svn.code.sf.net/p/freetzlinux/code@892 fc5b469a-3574-4ed2-a3ea-762ab7856fab

+++ trunk/speed-to-fritz/start-freetz.sh
@@ -5,6 +5,7 @@
 #so the firmware-versions in use will be downloaded and can be used by freetz later without copying them by hand.
 #---------------------
 # dont change variables
+DO_COPY="y" # just for testing
 export HOMEDIR="`pwd`"
 #include variables from last run of speed-to-fritz
 . ./incl_var
@@ -130,11 +131,11 @@ while [ "$KEY" != "y" ]; do
  echo "To download a specific revision of the trunk you would need to set the number via menu option!"
  echo
  
-if [ "$FREETZREVISION" ]; then
- echo -n "   Execute: 'svn co $FREETZ_DL_LINK $FREETZ_DIR' -r $FREETZREVISION (y/n)? "; read -n 1 -s YESNO; echo
-else
- echo -n "   Execute: 'svn co $FREETZ_DL_LINK $FREETZ_DIR'  (y/n)? "; read -n 1 -s YESNO; echo
-fi
+ if [ "$FREETZREVISION" ]; then
+  echo -n "   Execute: 'svn co $FREETZ_DL_LINK $FREETZ_DIR' -r $FREETZREVISION (y/n)? "; read -n 1 -s YESNO; echo
+ else
+  echo -n "   Execute: 'svn co $FREETZ_DL_LINK $FREETZ_DIR'  (y/n)? "; read -n 1 -s YESNO; echo
+ fi
  [ "$YESNO" = "y" ] || [ "$YESNO" = "n" ] &&  KEY="y"
  [ "$KEY" = "x" ] && echo "wrong key!"
  if [ "$YESNO" = "y" ]; then 
@@ -143,15 +144,18 @@ fi
   rm -fdr ./$FREETZ_DIR/toolchain/Config.in
   rm -fdr ./$FREETZ_DIR/patches/196-usbstorage.sh
   rm -fdr ./$FREETZ_DIR/patches/220-assistant.sh
-  rm -fdr ./$FREETZ_DIR//make/linux/kernel.mk
+  rm -fdr ./$FREETZ_DIR/make/linux/kernel.mk
   rm -fdr ./$FREETZ_DIR/fwmod
+  rm -fdr ./$FREETZ_DIR/patches/722A
+  rm -fdr ./$FREETZ_DIR/make/linux/patches/2.6.19.2/722A_04.76
+  rm -fdr ./$FREETZ_DIR/make/linux/Config.iks-16mb_26.722A_04.76
   echo "Looking for new freetz version, wait ..."
-    if [ "$FREETZREVISION" ]; then
+  if [ "$FREETZREVISION" ]; then
 	svn co $FREETZ_DL_LINK $FREETZ_DIR -r $FREETZREVISION
 	export SVN_VERSION="$FREETZREVISION"
-    else
+  else
 	svn co $FREETZ_DL_LINK $FREETZ_DIR
-    fi
+  fi
   #  export Revision=`svn co http://svn.freetz.org/trunk $FREETZ_DIR | grep 'Checked out revision' | tr -d '[:alpha:]' | tr -d '.'| tr -d ' '`
  fi
  [ -d "$FREETZ_DIR" ] || echo "freetz directory not found in the right place, should be in the same folder as speed-to-fritz!"
@@ -173,17 +177,7 @@ DL_DIR_ABS=./$FREETZ_DIR/$DL_DIR
 [ -d "$DL_DIR_ABS" ] || mkdir  "$DL_DIR_ABS"
 [ -d "$DL_DIR_ABS/fw" ] || mkdir  "$DL_DIR_ABS/fw"
 echo
-#echo -n "present directory is: "
-#pwd
-#echo "--->"
-#ls | grep "$FREETZ_DIR"
-#ls | grep 'speed-to-fritz'
 cd $FREETZ_DIR
-#echo
-#echo -n "present directory is: "
-#pwd
-#echo "--->"
-# ls
 echo "--Image files present in '$DL_DIR_ABS/fw':"
 ls -l $DL_DIR_ABS/fw/*.image 2>&1 | grep -v 'No such file' 
 echo
@@ -204,14 +198,19 @@ while [ "$KEY" != "y" ]; do
  echo -n "   Invoke 'make menuconfig' now? (y/n)? "; read -n 1 -s YESNO; echo
  [ "$YESNO" = "n" ] || [ "$YESNO" = "y" ] &&  KEY="y"
  [ "$KEY" = "x" ] && echo "wrong key!"
+ # menuconfig qestion -->
  if [ "$YESNO" = "y" ]; then
- #include variables from last run of speed-to-fritz
-. ../speed-to-fritz/incl_var
-
+  #include variables from last run of speed-to-fritz
+  . ../speed-to-fritz/incl_var
   echo "Be patient ..."
+  #7390 or W722-->
+  if [ "$FBMOD" = "7390" ] || [ "$SPNUM" = "722" ] ; then
+	# only W722 Config files
 	PATCH="$HOMEDIR/w722/W722_Config.diff"
-	patch -d . -p0 -N --no-backup-if-mismatch < "$PATCH" 2>&1
-
+	# add all W722 files 
+	[ "$DO_COPY" != "y" ] && PATCH="$HOMEDIR/w722/W722V_Type_A_files.diff"
+  	patch -d . -p0 -N --no-backup-if-mismatch < "$PATCH" 2>&1
+  fi  # 7390 or W722<--
   #preset freetz to the same type as speed-to-fritz
   sed -i -e "/Hardwaretype/,/---------/{/.*/d}" "./Config.in" 2> /dev/null
   sed -i -e '/General/a\
@@ -224,11 +223,11 @@ comment "----------------------------------------"' "./Config.in" 2> /dev/null
   sed -i -e "/config FREETZ_TYPE_SPEEDPORT_W501V/,/select FREETZ_DL_OVERRIDE/{/.*/d}" "./Config.in" 2> /dev/null
   sed -i -e '/default FREETZ_TYPE_FON_7150/d' "./Config.in" 2> /dev/null
               
-#echo "LABORTYPE: $TYPE_LABOR_TYPE  LANGUAGE:$avm_Lang  FBMOD:${FBMOD}"
-#sleep 5
-    [ "$avm_Lang" = "en" ] && sed -i -e "s/default FREETZ_TYPE_LANG_DE/default FREETZ_TYPE_LANG_EN/" "./Config.in" 2> /dev/null
-    [ "$avm_Lang" = "de" ] && sed -i -e "s/default FREETZ_TYPE_LANG_EN/default FREETZ_TYPE_LANG_DE/" "./Config.in" 2> /dev/null
-   sed -i -e '/FREETZ_TYPE/d' "./.config" 2> /dev/null
+  #echo "LABORTYPE: $TYPE_LABOR_TYPE  LANGUAGE:$avm_Lang  FBMOD:${FBMOD}"
+  #sleep 5
+  [ "$avm_Lang" = "en" ] && sed -i -e "s/default FREETZ_TYPE_LANG_DE/default FREETZ_TYPE_LANG_EN/" "./Config.in" 2> /dev/null
+  [ "$avm_Lang" = "de" ] && sed -i -e "s/default FREETZ_TYPE_LANG_EN/default FREETZ_TYPE_LANG_DE/" "./Config.in" 2> /dev/null
+  sed -i -e '/FREETZ_TYPE/d' "./.config" 2> /dev/null
   #[ "$avm_Lang" = "de" ] && sed -i -e 's/# FREETZ_TYPE_LANG_DE.*/FREETZ_TYPE_LANG_DE=y/' "./.config" 2> /dev/null
   #[ "$avm_Lang" = "de" ] && sed -i -e 's/FREETZ_TYPE_LANG_EN=.*/# FREETZ_TYPE_LANG_DE is not set/' "./.config" 2> /dev/null
 
@@ -243,11 +242,11 @@ comment "----------------------------------------"' "./Config.in" 2> /dev/null
   fi
   #7390 -->
   if [ "$FBMOD" = "7390" ] || [ "$SPNUM" = "722" ] ; then
-    echo "---------------------------------------------"
-    echo "FREETZ_TYPE_FON_WLAN_7390=y" >> "./.config" 2> /dev/null
-    #done ins in trunk now 4835
-    grep -q " -be" "./fwmod" && echo -e "\033[31mBig endianness option found!\033[0m"
-    if [ "$SPNUM" = "722" ] ; then
+   echo "---------------------------------------------"
+   echo "FREETZ_TYPE_FON_WLAN_7390=y" >> "./.config" 2> /dev/null
+   #done ins in trunk now 4835
+   grep -q " -be" "./fwmod" && echo -e "\033[31mBig endianness option found!\033[0m"
+   if [ "$SPNUM" = "722" ] ; then
     DISABLE_MODUL_AND_KERNEL_BILDING="n"
     if [ "$DISABLE_MODUL_AND_KERNEL_BILDING" = "y" ] ; then
      #--> diabele building of modules and kernel
@@ -256,23 +255,26 @@ comment "----------------------------------------"' "./Config.in" 2> /dev/null
      # disable error display if modules dir is not present
      sed -i -e 's|cd "${KERNEL_REP_DIR}/modules-${FREETZ_KERNEL_LAYOUT}-${FREETZ_KERNEL_REF}-${FREETZ_AVM_VERSION_STRING}"|cd "${KERNEL_REP_DIR}/modules-${FREETZ_KERNEL_LAYOUT}-${FREETZ_KERNEL_REF}-${FREETZ_AVM_VERSION_STRING}" 2>/dev/null|' "./fwmod"
      # preselect W722
+    fi #<-- 
+     # freetz revision 4844 
      echo "FREETZ_TYPE_SPEEDPORT_722A=y" >> "./.config" 2> /dev/null
-    fi
-     #<-- 
-        # copy patches to be usable with 7390 Source and w722 firmware
-	rm ./patches/7390/140-rc.S-no_avm_exit.patch
-	rm ./patches/7390/290-replace_websrv-remove_igdd.patch
-	rm ./patches/7390/301-remove_smbd.patch
-	rm ./patches/7390/de/*
-	rm ./patches/cond/usbstorage_7390.patch
-	cp $HOMEDIR/w722/patches/722A/*.patch ./patches/7390
-	cp $HOMEDIR/w722/patches/722A/de/* ./patches/7390/de
-	cp $HOMEDIR/w722/patches/722A/cond/* ./patches/cond
-	cp -fdrp $HOMEDIR/w722/patches/722A --target-directory="./patches"
-	cp -fdrp $HOMEDIR/w722/make --target-directory="."
-    fi
-    # add tcom as HTML directory as well, if a original tcom firmware is in use we only have tcom
-    sed -i -e "s|for i in all avme/en avme|for i in all avme/en avme tcom|" "./fwmod"
+     if [ "$DO_COPY" == "y" ]; then
+     # copy patches to be usable with 7390 Source and w722 firmware
+     	rm ./patches/7390/140-rc.S-no_avm_exit.patch
+     	rm ./patches/7390/290-replace_websrv-remove_igdd.patch
+     	rm ./patches/7390/301-remove_smbd.patch
+     	rm ./patches/7390/de/*
+     	rm ./patches/cond/usbstorage_7390.patch
+     	cp $HOMEDIR/w722/patches/722A/*.patch ./patches/7390
+     	cp $HOMEDIR/w722/patches/722A/de/* ./patches/7390/de
+     	cp $HOMEDIR/w722/patches/722A/cond/* ./patches/cond
+     	cp -fdrp $HOMEDIR/w722/patches/722A --target-directory="./patches"
+     	cp -fdrp $HOMEDIR/w722/make --target-directory="."
+     #cp -fdrpv $HOMEDIR/w722/make/linux/Config.iks-16mb_26.722A_04.76 ./make/linux/Config.iks-16mb_26.722A_04.76
+     fi
+   fi # <-- 722
+   # add tcom as HTML directory as well, if a original tcom firmware is in use we only have tcom
+   sed -i -e "s|for i in all avme/en avme|for i in all avme/en avme tcom|" "./fwmod"
   # 7390 <--
   else
     [ "$SPNUM" = "500" ] && echo "FREETZ_TYPE_WLAN_${FBMOD}=y" >> "./.config" 2> /dev/null
